Improved readability (I hope) ...
Main changes are added begin/end member functions to the Range class (formerly struct), the chart is now a simple vector instead of a "vector of vectors", which included helper functions to map the distance between two locations to the chart index. The arguments to std::inner_product were extracted for improved readability, which also massively improved the formatting.
Also, I added comments where I felt some explanation would be useful, if I was to look at this code again in a couple months/years.