Parsing is now done at compile-time instead of runtime
It bugged me that the bitset that represented whether or not a location was visited along the path, had to be fixed with some arbitrary hardcoded number to allow for a maximum of locations. Now, with compile-time parsing, the size of bitset will be automatically set depending on the input.